[dominion] replace menu with search content
